Move `Rejection` alongside the types that use it.
